# pkg/models/response.go
package models

import (
	"github.com/gin-gonic/gin"
)

type ErrorResponse struct {
	Error string `json:"error"`
}

type StatusResponse struct {
	Status string `json:"status"`
}

type CreateResponse struct {
	ID int `json:"id"`
}

type AccessTokenResponse struct {
	AccessToken string `json:"accessToken"`
}

func NewErrorResponse(c *gin.Context, statusCode int, message string) {
	c.AbortWithStatusJSON(statusCode, ErrorResponse{message})
}